% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Population structure
A structured object to describe population}
\usage{
population(identifiant, taille_ini, taux_rep, capacite)
}
\arguments{
\item{identifiant}{Valeur numérique identifiant l'espèce}

\item{taille_ini}{Taille initiale de la population en question}

\item{taux_rep}{Taux de reproduction maximum de la population}

\item{capacite}{Capacite biotique du milieu}
}
\value{
Une liste structuré comprenant les paramètres de la population mentionnés ci-dessus
}
\description{
Population structure
A structured object to describe population
}
\examples{
population(34, 50, 0.8, 5000)
}
