% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{print.Population}
\alias{print.Population}
\title{Print population}
\usage{
\method{print}{Population}(x)
}
\arguments{
\item{x}{Un environnement associé à la population}
}
\value{
Taille de la population par identifiant pour chaque identifiant
}
\description{
Print population
}
\examples{
print.Population(pop)

}
